+++ test/number_extension_test.dart
@@ -6,10 +6,10 @@ void main() {
   test('adds one to input values', () {
     final numberSystem = NumberExtension();
     expect(numberSystem.internationSystem(100000), '100 K');
-    expect(numberSystem.internationSystem(1000000), '1 M');
-    expect(numberSystem.internationSystem(10000000000000), '10 T');
-    expect(numberSystem.indianSystem(100000), '1 Lac');
-    expect(numberSystem.indianSystem(100000000), '10 Crore');
-    expect(numberSystem.indianSystem(1000000000), '1 Arab');
+    expect(numberSystem.internationSystem(1000000), '1.0 M');
+    expect(numberSystem.internationSystem(10000000000000), '10.0 T');
+    expect(numberSystem.indianSystem(100000), '1.0 Lac');
+    expect(numberSystem.indianSystem(100000000), '10.0 Crore');
+    expect(numberSystem.indianSystem(1000000000), '1.0 Arab');
   });
 }